requirements.txt
opencv-python
numpy
rembg
pillow
onnxruntime
python-multipart
uvicorn
fastapi
aiofiles
pydantic
fastapi-utils
